% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun.R
\name{test_fun}
\alias{test_fun}
\title{Function demonstrating segfault with flang}
\usage{
test_fun(N)
}
\arguments{
\item{N}{Integer number}
}
\value{
nothing
}
\description{
Minimal example reproducing an issue
   in package sequoia
}
\examples{
test_fun(510^2*8)   # N<=510 runs OK
test_fun(511^2*8)   # N>=511 segfault

}
